Added estimatedDecodingFrequency as debug property of RemotePlayMonitor.
Controls how hard the server will throttle the data it sends to the client. (1 is the most throttling, 255 the least)

+++ client/SimulCasterClient/src/Application.cpp
@@ -845,7 +845,7 @@ void Application::OnVideoStreamChanged(const avs::SetupCommand &setupCommand,avs
 
     handshake.framerate=60;
 	handshake.udpBufferSize=mNetworkSource.getSystemBufferSize();
-	handshake.maxBandwidth=handshake.udpBufferSize*(size_t)handshake.framerate;
+	handshake.maxBandwidth = handshake.udpBufferSize * static_cast<uint32_t>(handshake.framerate);
 }
 
 void Application::OnVideoStreamClosed()
